requirements.txt
python-dotenv==1.0.0
Django==4.2.8
djangorestframework==3.14.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
coreapi==2.3.3
coreschema==0.0.4
django-anymail==10.2